working to restructure Training neighborlist.
