+++ modelforge/potential/neighbors.py
@@ -2,6 +2,7 @@ import torch
 from loguru import logger as log
 from modelforge.potential.models import PairlistData
 from modelforge.dataset.dataset import NNPInputTuple
+from modelforge.potential.models import Pairlist
 
 
 class OrthogonalDisplacementFunction(torch.nn.Module):
@@ -440,3 +441,115 @@ class NeighborlistVerletNsq(torch.nn.Module):
                 d_ij=d_ij_full,
                 r_ij=r_ij_full,
             )
+
+
+class NeighborForlistTraining(torch.nn.Module):
+
+    def __init__(self, cutoff: float, only_unique_pairs: bool = False):
+        """
+        Initialize the ComputeInteractingAtomPairs module.
+
+        Parameters
+        ----------
+        cutoff : float
+            The cutoff distance for neighbor list calculations.
+        only_unique_pairs : bool, optional
+            If True, only unique pairs are returned (default is False).
+        """
+
+        super().__init__()
+        from .models import Neighborlist
+
+        self.only_unique_pairs = only_unique_pairs
+        self.pairlist = Pairlist(cutoff, only_unique_pairs)
+        self.register_buffer("cutoff", torch.tensor(cutoff))
+
+    def _calculate_interacting_pairs(
+        self,
+        positions: torch.Tensor,
+        atomic_subsystem_indices: torch.Tensor,
+        pair_indices: Optional[torch.Tensor] = None,
+    ) -> PairlistData:
+        """
+        Compute the neighbor list considering a cutoff distance.
+
+        Parameters
+        ----------
+        positions : torch.Tensor
+            Atom positions. Shape: [nr_systems, nr_atoms, 3].
+        atomic_subsystem_indices : torch.Tensor
+            Indices identifying atoms in subsystems. Shape: [nr_atoms].
+        pair_indices : Optional[torch.Tensor]
+            Precomputed pair indices. If None, will compute pair indices.
+
+        Returns
+        -------
+        PairListOutputs
+            A dataclass containing 'pair_indices', 'd_ij' (distances), and 'r_ij' (displacement vectors).
+        """
+
+        if pair_indices is None:
+            pair_indices = self.enumerate_all_pairs(
+                atomic_subsystem_indices,
+            )
+
+        r_ij = self.calculate_r_ij(pair_indices, positions)
+        d_ij = self.calculate_d_ij(r_ij)
+
+        in_cutoff = (d_ij <= self.cutoff).squeeze()
+        # Get the atom indices within the cutoff
+        pair_indices_within_cutoff = pair_indices[:, in_cutoff]
+
+        return PairlistData(
+            pair_indices=pair_indices_within_cutoff,
+            d_ij=d_ij[in_cutoff],
+            r_ij=r_ij[in_cutoff],
+        )
+
+    def forward(self, data: Union[NNPInput, NamedTuple]) -> PairlistData:
+        """
+        Compute the pair list, distances, and displacement vectors for the given
+        input data.
+
+        Parameters
+        ----------
+        data : Union[NNPInput, NamedTuple]
+            Input data containing atomic numbers, positions, and subsystem
+            indices.
+
+        Returns
+        -------
+        PairlistData
+            A namedtuple containing the pair indices, distances, and
+            displacement vectors.
+        """
+        # ---------------------------
+        # general input manipulation
+        positions = data.positions
+        atomic_subsystem_indices = data.atomic_subsystem_indices
+        # calculate pairlist if none is provided
+        if data.pair_list is None:
+            pair_list = self.pairlist.enumerate_all_pairs(atomic_subsystem_indices)
+
+        else:
+            # pairlist is provided, remove redundant pairs if requested
+            if self.only_unique_pairs:
+                i_indices = data.pair_list[0]
+                j_indices = data.pair_list[1]
+                unique_pairs_mask = i_indices < j_indices
+                i_final_pairs = i_indices[unique_pairs_mask]
+                j_final_pairs = j_indices[unique_pairs_mask]
+                pair_list = torch.stack((i_final_pairs, j_final_pairs))
+            else:
+                pair_list = data.pair_list
+            # only calculate d_ij and r_ij
+            pairlist_output = self.calculate_distances_and_pairlist(
+                positions=positions,
+                atomic_subsystem_indices=atomic_subsystem_indices,
+                pair_indices=pair_list.to(torch.int64),
+            )
+
+        return pairlist_output
+
+
+#
